# backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Eki backend HTTP API",
    "version": "1.0.0",
    "description": "HTTP contract for the 39 existing handlers and six compatible v2 aliases from PR #197. Firebase SDK listeners are documented separately in docs/api/HTTP_CONTRACT.md; they are not HTTP operations. Legacy URLs remain supported; no retirement date is scheduled. REST naming alone promises no performance improvement."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development. Production clients must configure the deployed HTTPS backend origin."
    }
  ],
...
  "paths": {
...
    "/api/routes/{routeId}": {
      "put": {
        "operationId": "saveRoute",
        "tags": [
          "Routes"
        ],
        "summary": "Save a versioned route with durable operation reconciliation",
        "description": "create requires expectedVersion=0; edit requires current configVersion. Unique ordered stop IDs and active-ride checks rechecked at commit. Metadata-only edits reuse valid geometry; configVersion advances every edit and geometryVersion only for geometry changes. 202 is processing, never confirmed success.",
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteSaveInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteSaved"
                }
              }
            }
          },
          "202": {
            "description": "Accepted or still processing; consult operation-specific semantics.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteProcessing"
                }
              }
            }
          },
          "400": {
            "description": "Invalid identifiers, payload, query, or JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, revoked, disabled, or malformed credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated caller lacks the required role, assignment, proximity, or membership.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Lifecycle, version, assignment, or idempotency conflict; reconcile before retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Raw JSON request exceeds the configured body limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded; obey Retry-After when supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry when emitted by the rejecting limiter; not supplied by every upstream failure.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal operation failure; details are redacted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Upstream route/places service failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable, busy authentication verifier, or missing safe server configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Upstream deadline exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-auth-class": "admin",
        "x-body-limit-bytes": 16384,
        "x-rate-limit": "Global 200 requests/verified UID or normalized IP/minute, divided by RATE_LIMIT_SHARD_FACTOR. The edge WAF/LB is the global fleet cap. Route mount: 10 requests/IP/minute divided by shard factor, including geometry reads and deletes. GET save-operation reconciliation skips this dedicated limiter.",
        "x-cache-policy": "No explicit Cache-Control header from this handler. Authenticated clients must not use shared response caches. This does not imply private server-side caches are absent.",
        "x-timeout-policy": "HTTP request-body receipt timeout 15000 ms (not a handler execution deadline); header timeout 70000 ms. No general handler deadline; clients use bounded timeouts and reconcile uncertain writes. Each Google Routes chunk has a 10000 ms deadline. Up to 100 ordered points are chunked into forward and reverse routing calls; no whole-handler deadline. Browser save/reconciliation budget is 30000 ms; durable save lease is 30000 ms.",
        "x-retry-policy": "Do not blindly retry 400/401/403/409/422. On 429 obey Retry-After when present. Retry transient failures with bounded exponential backoff and jitter only when the operation idempotency policy permits.",
        "x-idempotency": "Stable saveId binds the normalized route ID, mode, expectedVersion, metadata and ordered stops. Same key/payload replays stored result or failure. Changed payload conflicts. Unknown outcome/202: GET the same save-operation resource, then retry the exact same payload/key if appropriate."
      },
      "delete": {
        "operationId": "deleteRoute",
        "tags": [
          "Routes"
        ],
        "summary": "Delete a route after assignment and ride safety checks",
        "description": "Delete a route after assignment and ride safety checks",
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "parameters": [
          {
            "name": "routeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{1,128}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid identifiers, payload, query, or JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, revoked, disabled, or malformed credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated caller lacks the required role, assignment, proximity, or membership.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Lifecycle, version, assignment, or idempotency conflict; reconcile before retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Raw JSON request exceeds the configured body limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded; obey Retry-After when supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry when emitted by the rejecting limiter; not supplied by every upstream failure.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal operation failure; details are redacted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable, busy authentication verifier, or missing safe server configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-auth-class": "admin",
        "x-body-limit-bytes": 16384,
        "x-rate-limit": "Global 200 requests/verified UID or normalized IP/minute, divided by RATE_LIMIT_SHARD_FACTOR. The edge WAF/LB is the global fleet cap. Route mount: 10 requests/IP/minute divided by shard factor, including geometry reads and deletes. GET save-operation reconciliation skips this dedicated limiter.",
        "x-cache-policy": "No explicit Cache-Control header from this handler. Authenticated clients must not use shared response caches. This does not imply private server-side caches are absent.",
        "x-timeout-policy": "HTTP request-body receipt timeout 15000 ms (not a handler execution deadline); header timeout 70000 ms. No general handler deadline; clients use bounded timeouts and reconcile uncertain writes.",
        "x-retry-policy": "Do not blindly retry 400/401/403/409/422. On 429 obey Retry-After when present. Retry transient failures with bounded exponential backoff and jitter only when the operation idempotency policy permits.",
        "x-idempotency": "No replay key. Reconcile unknown outcomes; already absent returns 404."
      }
    },
...
    "/api/v2/routes": {
      "get": {
        "operationId": "listRoutes",
        "tags": [
          "Planning"
        ],
        "summary": "Read bounded route collection projection",
        "description": "Same bounded Firestore projection as legacy routes-list: at most 250 records and no geometry. It queries Firestore per request.",
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid identifiers, payload, query, or JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, revoked, disabled, or malformed credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Raw JSON request exceeds the configured body limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded; obey Retry-After when supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry when emitted by the rejecting limiter; not supplied by every upstream failure.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal operation failure; details are redacted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable, busy authentication verifier, or missing safe server configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-auth-class": "user",
        "x-body-limit-bytes": 16384,
        "x-rate-limit": "Global 200 requests/verified UID or normalized IP/minute, divided by RATE_LIMIT_SHARD_FACTOR. The edge WAF/LB is the global fleet cap.",
        "x-cache-policy": "Handler emits Cache-Control: no-store. Earlier middleware/parser rejections may occur before that header.",
        "x-timeout-policy": "HTTP request-body receipt timeout 15000 ms (not a handler execution deadline); header timeout 70000 ms. No general handler deadline; clients use bounded timeouts and reconcile uncertain writes.",
        "x-retry-policy": "Do not blindly retry 400/401/403/409/422. On 429 obey Retry-After when present. Retry transient failures with bounded exponential backoff and jitter only when the operation idempotency policy permits.",
        "x-idempotency": "Safe snapshot/read; no mutation replay key."
      }
    },
...
    "/api/settings": {
      "put": {
        "operationId": "putSettingsLegacy",
        "tags": [
          "Settings"
        ],
        "summary": "Apply a validated partial settings update",
        "description": "Apply a validated partial settings update",
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Saved"
                }
              }
            }
          },
          "400": {
            "description": "Invalid identifiers, payload, query, or JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, revoked, disabled, or malformed credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated caller lacks the required role, assignment, proximity, or membership.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Raw JSON request exceeds the configured body limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded; obey Retry-After when supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry when emitted by the rejecting limiter; not supplied by every upstream failure.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal operation failure; details are redacted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable, busy authentication verifier, or missing safe server configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-auth-class": "admin",
        "x-body-limit-bytes": 16384,
        "x-rate-limit": "Global 200 requests/verified UID or normalized IP/minute, divided by RATE_LIMIT_SHARD_FACTOR. The edge WAF/LB is the global fleet cap. Mount-level writeLimiter: 30 requests/identity/minute, divided by RATE_LIMIT_SHARD_FACTOR; this also covers GETs on these mounts.",
        "x-cache-policy": "Handler emits Cache-Control: no-store. Earlier middleware/parser rejections may occur before that header.",
        "x-timeout-policy": "HTTP request-body receipt timeout 15000 ms (not a handler execution deadline); header timeout 70000 ms. No general handler deadline; clients use bounded timeouts and reconcile uncertain writes.",
        "x-retry-policy": "Do not blindly retry 400/401/403/409/422. On 429 obey Retry-After when present. Retry transient failures with bounded exponential backoff and jitter only when the operation idempotency policy permits.",
        "x-idempotency": "Partial merge rather than full replacement; repeated values converge but server audit timestamp refreshes.",
        "x-compatibility-alias-of": "PATCH /api/v2/settings/global"
      }
    },
    "/api/v2/settings/global": {
      "patch": {
        "operationId": "patchSettings",
        "tags": [
          "Settings"
        ],
        "summary": "Apply a validated partial global settings update",
        "description": "One or more allowed bounded strings/announcementActive boolean, rejects unknown keys. Non-announcement strings must be nonblank. Shares validation and Firestore merge/audit logic with legacy PUT.",
        "security": [
          {
            "FirebaseBearer": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Saved"
                }
              }
            }
          },
          "400": {
            "description": "Invalid identifiers, payload, query, or JSON.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, revoked, disabled, or malformed credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated caller lacks the required role, assignment, proximity, or membership.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Raw JSON request exceeds the configured body limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded; obey Retry-After when supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds before retry when emitted by the rejecting limiter; not supplied by every upstream failure.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal operation failure; details are redacted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency unavailable, busy authentication verifier, or missing safe server configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-auth-class": "admin",
        "x-body-limit-bytes": 16384,
        "x-rate-limit": "Global 200 requests/verified UID or normalized IP/minute, divided by RATE_LIMIT_SHARD_FACTOR. The edge WAF/LB is the global fleet cap. Mount-level writeLimiter: 30 requests/identity/minute, divided by RATE_LIMIT_SHARD_FACTOR; this also covers GETs on these mounts.",
        "x-cache-policy": "Handler emits Cache-Control: no-store. Earlier middleware/parser rejections may occur before that header.",
        "x-timeout-policy": "HTTP request-body receipt timeout 15000 ms (not a handler execution deadline); header timeout 70000 ms. No general handler deadline; clients use bounded timeouts and reconcile uncertain writes.",
        "x-retry-policy": "Do not blindly retry 400/401/403/409/422. On 429 obey Retry-After when present. Retry transient failures with bounded exponential backoff and jitter only when the operation idempotency policy permits.",
        "x-idempotency": "Partial merge; unspecified fields survive. Repeats refresh server audit metadata."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "FirebaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token",
        "description": "Revocation-checked ID token. Role, assignment, membership, and privileged-account constraints are enforced per operation; a valid token alone does not grant admin rights."
      },
      "DeviceCredential": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Send the complete header value Device <per-device secret>. Never use Bearer or send this credential to the firmware artifact host."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "retryAfterMs": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "outcomeUnknown": {
            "type": "boolean"
          },
          "currentVersion": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": true
      },
...
      "Deleted": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "deleted"
        ],
        "additionalProperties": false
      },
...
      "RouteList": {
        "type": "object",
        "properties": {
          "routes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "pattern": "^[A-Za-z0-9_-]{1,128}$"
                },
                "name": {
                  "type": "string"
                },
                "color": {
                  "type": "string"
                },
                "stops": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Stop"
                  }
                }
              },
              "required": [
                "id",
                "name",
                "color",
                "stops"
              ],
              "additionalProperties": false
            },
            "minItems": 0,
            "maxItems": 250
          }
        },
        "required": [
          "routes"
        ],
        "additionalProperties": false
      },
...
      "SettingsPatch": {
        "type": "object",
        "properties": {
          "serviceStartTime": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1
          },
          "noBusesMessage": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          },
          "noBusesSubMessage": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1
          },
          "announcementText": {
            "type": "string",
            "maxLength": 500
          },
          "announcementActive": {
            "type": "boolean"
          }
        },
        "required": [],
        "additionalProperties": false,
        "minProperties": 1
      },
...
      "RouteSaveInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1
          },
          "color": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$"
          },
          "mode": {
            "type": "string",
            "enum": [
              "create",
              "edit"
            ]
          },
          "saveId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,128}$"
          },
          "expectedVersion": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteStopInput"
            },
            "minItems": 2,
            "maxItems": 100
          }
        },
        "required": [
          "name",
          "color",
          "mode",
          "saveId",
          "expectedVersion",
          "stops"
        ],
        "additionalProperties": true
      },
      "RouteSaved": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Geometry"
          },
          {
            "type": "object",
            "properties": {
              "saved": {
                "type": "boolean",
                "const": true
              },
              "status": {
                "const": "succeeded",
                "type": "string"
              },
              "saveId": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]{1,128}$"
              },
              "routeId": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]{1,128}$"
              },
              "configVersion": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "geometryVersion": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "geometryReused": {
                "type": "boolean"
              }
            },
            "required": [
              "saved",
              "status",
              "saveId",
              "routeId",
              "configVersion",
              "geometryVersion",
              "geometryReused"
            ],
            "additionalProperties": true
          }
        ]
      },
      "RouteProcessing": {
        "type": "object",
        "properties": {
          "status": {
            "const": "processing",
            "type": "string"
          },
          "saveId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{1,128}$"
          },
          "retryAfterMs": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "status",
          "saveId",
          "retryAfterMs"
        ],
        "additionalProperties": false
      },
...
    }
  }
}